@@ -1,6 +1,6 @@
 {
   "name": "demo",
-  "version": "0.6.3",
+  "version": "0.7.5",
   "private": true,
   "scripts": {
     "docusaurus": "docusaurus",
@@ -14,15 +14,15 @@
     "write-heading-ids": "docusaurus write-heading-ids"
   },
   "dependencies": {
-    "@docusaurus/core": "^2.0.0",
-    "@mdx-js/react": "^1.6.22",
+    "@docusaurus/core": "^3.6.0",
+    "@docusaurus/faster": "^3.6.0",
     "@svgr/webpack": "^5.5.0",
-    "clsx": "^1.1.1",
-    "docusaurus-preset-openapi": "^0.6.3",
+    "clsx": "^1.2.1",
+    "docusaurus-preset-openapi": "^0.7.5",
     "file-loader": "^6.2.0",
-    "prism-react-renderer": "^1.2.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "prism-react-renderer": "^2.1.0",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "url-loader": "^4.1.1"
   },
   "browserslist": {
