@@ -38,9 +38,9 @@
     "@babel/preset-typescript": "^7.16.0",
     "@testing-library/cypress": "^8.0.1",
     "@types/jest": "^27.0.2",
-    "@types/node": "^17.0.2",
-    "@typescript-eslint/eslint-plugin": "^4.0.0",
-    "@typescript-eslint/parser": "^4.0.0",
+    "@types/node": "^18.0.0",
+    "@typescript-eslint/eslint-plugin": "^6.0.0",
+    "@typescript-eslint/parser": "^6.0.0",
     "babel-eslint": "^10.0.0",
     "cross-env": "^7.0.3",
     "cypress": "^8.7.0",
@@ -66,7 +66,7 @@
     "start-server-and-test": "^1.14.0",
     "ts-jest": "^27.0.6",
     "ts-node": "^10.9.1",
-    "typescript": "^4.7.4"
+    "typescript": "^5.2.2"
   },
   "lint-staged": {
     "**/*": "prettier --write --ignore-unknown"
