@@ -1,7 +1,7 @@
 {
   "name": "docusaurus-preset-openapi",
   "description": "OpenAPI preset for Docusaurus.",
-  "version": "0.6.3",
+  "version": "0.7.5",
   "license": "MIT",
   "keywords": [
     "openapi",
@@ -28,19 +28,19 @@
     "watch": "tsc --watch"
   },
   "devDependencies": {
-    "@docusaurus/types": "^2.0.0"
+    "@docusaurus/types": "^3.6.0"
   },
   "dependencies": {
-    "@docusaurus/preset-classic": "^2.0.0",
-    "docusaurus-plugin-openapi": "^0.6.3",
-    "docusaurus-plugin-proxy": "^0.6.3",
-    "docusaurus-theme-openapi": "^0.6.3"
+    "@docusaurus/preset-classic": "^3.6.0",
+    "docusaurus-plugin-openapi": "^0.7.5",
+    "docusaurus-plugin-proxy": "^0.7.5",
+    "docusaurus-theme-openapi": "^0.7.5"
   },
   "peerDependencies": {
-    "react": "^16.8.4 || ^17.0.0",
-    "react-dom": "^16.8.4 || ^17.0.0"
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0"
   },
   "engines": {
-    "node": ">=14"
+    "node": ">=18"
   }
 }
