@@ -1,8 +1,10 @@
 // @ts-check
-// Note: type annotations allow type checking and IDEs autocompletion
+// `@type` JSDoc annotations allow editor autocompletion and type checking
+// (when paired with `@ts-check`).
+// There are various equivalent ways to declare your Docusaurus config.
+// See: https://docusaurus.io/docs/api/docusaurus-config
 
-const lightCodeTheme = require("prism-react-renderer/themes/github");
-const darkCodeTheme = require("prism-react-renderer/themes/dracula");
+import { themes as prismThemes } from "prism-react-renderer";
 
 /** @type {import('@docusaurus/types').Config} */
 const config = {
@@ -15,6 +17,18 @@ const config = {
   favicon: "img/favicon.ico",
   organizationName: "cloud-annotations", // Usually your GitHub org/user name.
   projectName: "docusaurus-openapi", // Usually your repo name.
+
+  markdown: {
+    mdx1Compat: {
+      comments: false,
+      admonitions: false,
+      headingIds: false,
+    },
+  },
+
+  future: {
+    experimental_faster: true,
+  },
 
   presets: [
     [
@@ -113,8 +127,8 @@ const config = {
         copyright: `Copyright © ${new Date().getFullYear()} Cloud Annotations. Built with Docusaurus.`,
       },
       prism: {
-        theme: lightCodeTheme,
-        darkTheme: darkCodeTheme,
+        theme: prismThemes.github,
+        darkTheme: prismThemes.dracula,
       },
       api: {
         authPersistance: "localStorage",
@@ -123,4 +137,4 @@ const config = {
     }),
 };
 
-module.exports = config;
+export default config;
