@@ -5,8 +5,10 @@
  * LICENSE file in the root directory of this source tree.
  * ========================================================================== */
 
-import path from "path";
+import { readFile } from "node:fs/promises";
+import { join, relative, resolve } from "node:path";
 
+import type { Options as MDXLoaderOptions } from "@docusaurus/mdx-loader";
 import type {
   LoadContext,
   Plugin,
@@ -23,8 +25,7 @@ import {
   Globby,
 } from "@docusaurus/utils";
 import chalk from "chalk";
-import fs from "fs-extra";
-import { Configuration } from "webpack";
+import type { Configuration } from "webpack";
 
 import { DocEnv, processDocMetadata } from "./docs/docs";
 import { createApiPageMD, createInfoPageMD } from "./markdown";
@@ -37,23 +38,23 @@ export default function pluginOpenAPI(
   context: LoadContext,
   options: PluginOptions
 ): Plugin<LoadedContent> {
-  const { baseUrl, generatedFilesDir, siteDir } = context;
+  const { baseUrl, generatedFilesDir, siteDir, siteConfig } = context;
 
   const pluginId = options.id ?? DEFAULT_PLUGIN_ID;
 
-  const pluginDataDirRoot = path.join(
+  const pluginDataDirRoot = join(
     generatedFilesDir,
     "docusaurus-plugin-openapi"
   );
 
-  const dataDir = path.join(pluginDataDirRoot, pluginId);
+  const dataDir = join(pluginDataDirRoot, pluginId);
 
   const aliasedSource = (source: string) =>
-    `~api/${posixPath(path.relative(pluginDataDirRoot, source))}`;
+    `~api/${posixPath(relative(pluginDataDirRoot, source))}`;
 
   const contentPath = isURL(options.path)
     ? options.path
-    : path.resolve(context.siteDir, options.path);
+    : resolve(context.siteDir, options.path);
 
   return {
     name: "docusaurus-plugin-openapi",
@@ -70,8 +71,8 @@ export default function pluginOpenAPI(
         /** E.g. "api/plugins/myDoc.mdx" */
         relativeSource: string
       ): Promise<MdxPageMetadata> {
-        const source = path.join(contentPath, relativeSource);
-        const content = await fs.readFile(source, "utf-8");
+        const source = join(contentPath, relativeSource);
+        const content = await readFile(source, "utf-8");
 
         return {
           type: "mdx",
@@ -200,7 +201,7 @@ export default function pluginOpenAPI(
           exact: true,
           modules: {
             // TODO: "-content" should be inside hash to prevent name too long errors.
-            content: path.join(dataDir, `${docuHash(pageId)}-content.mdx`),
+            content: join(dataDir, `${docuHash(pageId)}-content.mdx`),
           },
           sidebar: sidebarName,
         };
@@ -237,6 +238,7 @@ export default function pluginOpenAPI(
         modules: {
           apiMetadata: aliasedSource(apiBaseMetadataPath),
         },
+        priority: 999999,
       });
 
       return;
@@ -244,12 +246,13 @@ export default function pluginOpenAPI(
 
     configureWebpack(
       _config: Configuration,
-      isServer: boolean,
-      { getJSLoader }: ConfigureWebpackUtils
+      _isServer: boolean,
+      _configureWebpackUtils: ConfigureWebpackUtils
     ) {
       const {
         rehypePlugins,
         remarkPlugins,
+        recmaPlugins,
         beforeDefaultRehypePlugins,
         beforeDefaultRemarkPlugins,
       } = options;
@@ -266,16 +269,20 @@ export default function pluginOpenAPI(
               test: /(\.mdx?)$/,
               include: [dataDir, contentPath].map(addTrailingPathSeparator),
               use: [
-                getJSLoader({ isServer }),
                 {
                   loader: require.resolve("@docusaurus/mdx-loader"),
                   options: {
+                    admonitions: options.admonitions,
                     remarkPlugins,
                     rehypePlugins,
+                    recmaPlugins,
                     beforeDefaultRehypePlugins,
                     beforeDefaultRemarkPlugins,
-                    // Docusaurus 2.2.0 has a regression that requires this option to be set.
-                    markdownConfig: {},
+                    staticDirs: siteConfig.staticDirectories.map((dir) =>
+                      resolve(siteDir, dir)
+                    ),
+                    siteDir,
+                    markdownConfig: siteConfig.markdown ?? { mdx1Compat: {} },
                     metadataPath: (mdxPath: string) => {
                       if (mdxPath.startsWith(dataDir)) {
                         // The MDX file already lives in `dataDir`: this is an OpenAPI MDX
@@ -283,13 +290,10 @@ export default function pluginOpenAPI(
                       } else {
                         // Standard resolution
                         const aliasedSource = aliasedSitePath(mdxPath, siteDir);
-                        return path.join(
-                          dataDir,
-                          `${docuHash(aliasedSource)}.json`
-                        );
+                        return join(dataDir, `${docuHash(aliasedSource)}.json`);
                       }
                     },
-                  },
+                  } satisfies MDXLoaderOptions,
                 },
               ].filter(Boolean),
             },
