@@ -32,8 +32,21 @@ export function guard<T>(
 }
 
 export function render(children: Children): string {
-  if (Array.isArray(children)) {
-    return children.filter((c) => c !== undefined).join("");
+  const res = Array.isArray(children)
+    ? children.filter((c) => c !== undefined).join("\n")
+    : children ?? "";
+
+  const isMultiline = res.split("\n").length > 1;
+
+  // It is not possible to wrap “blocks” if text and tags are on the same line,
+  // but the corresponding tags are on different lines. This can accidentally
+  // happen if the rendered item has multiple lines. To be safe, we pad with
+  // newlines.
+  //
+  // See: https://mdxjs.com/migrating/v2/#jsx
+  if (isMultiline) {
+    return "\n" + res + "\n";
   }
-  return children ?? "";
+
+  return res;
 }
