@@ -9,10 +9,6 @@ import React from "react";
 
 import { RequestBodyObject } from "docusaurus-plugin-openapi/src/openapi/types";
 
-import ContentType from "../ContentType";
-import FormMultiSelect from "../FormMultiSelect";
-import FormSelect from "../FormSelect";
-import { useTypedDispatch, useTypedSelector } from "../hooks";
 import FormFileUpload from "./../FormFileUpload";
 import FormItem from "./../FormItem";
 import FormTextInput from "./../FormTextInput";
@@ -26,6 +22,10 @@ import {
   setArrayFormBody,
   setStringRawBody,
 } from "./slice";
+import ContentType from "../ContentType";
+import FormMultiSelect from "../FormMultiSelect";
+import FormSelect from "../FormSelect";
+import { useTypedDispatch, useTypedSelector } from "../hooks";
 
 interface Props {
   jsonRequestBodyExample: string;
